requirements.txt
psycopg2-binary
pytest
requests
watchdog
